requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Voice: ASR
faster-whisper>=1.0.0

# Voice: TTS
elevenlabs>=1.0.0

# Agent (for Backend integration)
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# MCP / Playwright (for agent team)
playwright>=1.40.0
